website: introduce the multi_a fragment, available for Tapscript
multi_a() has the same properties than multi() but 'n' (since all keys
may need to be dissatisfied, the top stack item of a satisfaction may be
the empty vectore). It could also be 'o' when it contains a single key..
But a `multi_a(1,key)` can always be subsituted by a `pk(key)`.